Add files via upload
The following code applies the models of CitizenLab, Lxyuan, and Sentic to a list of tweets, taking a list of tweets as input and returning binary columns. Each model has its own column, and if it has a value of 1, it indicates a positive sentiment, while 0 indicates a negative sentiment.